+++ cmd/exo/cmd/api.go
@@ -3,18 +3,14 @@ package cmd
 import (
 	"encoding/json"
 	"fmt"
-	"io"
 	"log"
 	"net"
 	"os"
 	"reflect"
 	"strings"
-	"syscall"
-	"text/tabwriter"
 
 	"github.com/exoscale/egoscale"
 	"github.com/spf13/cobra"
-	"golang.org/x/crypto/ssh/terminal"
 )
 
 // apiCmd represents the api command
@@ -270,202 +266,5 @@ func buildFlags(method egoscale.Command, cmd *cobra.Command) {
 		default:
 			log.Printf("[SKIP] Type of %s is not supported! %v", field.Name, val.Kind())
 		}
-	}
-}
-
-type cmd struct {
-	command egoscale.Command
-	hidden  bool
-}
-
-var methods = map[string][]cmd{
-	"network": {
-		{&egoscale.CreateNetwork{}, false},
-		{&egoscale.DeleteNetwork{}, false},
-		{&egoscale.ListNetworkOfferings{}, false},
-		{&egoscale.ListNetworks{}, false},
-		{&egoscale.RestartNetwork{}, true},
-		{&egoscale.UpdateNetwork{}, false},
-	},
-	"virtual-machine": {
-		{&egoscale.AddNicToVirtualMachine{}, false},
-		{&egoscale.ChangeServiceForVirtualMachine{}, false},
-		{&egoscale.DeployVirtualMachine{}, false},
-		{&egoscale.DestroyVirtualMachine{}, false},
-		{&egoscale.ExpungeVirtualMachine{}, false},
-		{&egoscale.GetVMPassword{}, false},
-		{&egoscale.GetVirtualMachineUserData{}, false},
-		{&egoscale.ListVirtualMachines{}, false},
-		{&egoscale.MigrateVirtualMachine{}, true},
-		{&egoscale.RebootVirtualMachine{}, false},
-		{&egoscale.RecoverVirtualMachine{}, false},
-		{&egoscale.RemoveNicFromVirtualMachine{}, false},
-		{&egoscale.ResetPasswordForVirtualMachine{}, false},
-		{&egoscale.RestoreVirtualMachine{}, false},
-		{&egoscale.ScaleVirtualMachine{}, false},
-		{&egoscale.StartVirtualMachine{}, false},
-		{&egoscale.StopVirtualMachine{}, false},
-		{&egoscale.UpdateDefaultNicForVirtualMachine{}, true},
-		{&egoscale.UpdateVirtualMachine{}, false},
-	},
-	"volume": {
-		{&egoscale.ListVolumes{}, false},
-		{&egoscale.ResizeVolume{}, false},
-	},
-	"template": {
-		{&egoscale.CopyTemplate{}, true},
-		{&egoscale.CreateTemplate{}, true},
-		{&egoscale.ListTemplates{}, false},
-		{&egoscale.PrepareTemplate{}, true},
-		{&egoscale.RegisterTemplate{}, true},
-		{&egoscale.ListOSCategories{}, true},
-	},
-	"account": {
-		{&egoscale.EnableAccount{}, true},
-		{&egoscale.DisableAccount{}, true},
-		{&egoscale.ListAccounts{}, false},
-	},
-	"zone": {
-		{&egoscale.ListZones{}, false},
-	},
-	"snapshot": {
-		{&egoscale.CreateSnapshot{}, false},
-		{&egoscale.DeleteSnapshot{}, false},
-		{&egoscale.ListSnapshots{}, false},
-		{&egoscale.RevertSnapshot{}, false},
-	},
-	"user": {
-		{&egoscale.CreateUser{}, true},
-		{&egoscale.DeleteUser{}, true},
-		//{&egoscale.DisableUser{}, true},
-		//{&egoscale.GetUser{}, true},
-		{&egoscale.UpdateUser{}, true},
-		{&egoscale.ListUsers{}, false},
-		{&egoscale.RegisterUserKeys{}, false},
-	},
-	"security-group": {
-		{&egoscale.AuthorizeSecurityGroupEgress{}, false},
-		{&egoscale.AuthorizeSecurityGroupIngress{}, false},
-		{&egoscale.CreateSecurityGroup{}, false},
-		{&egoscale.DeleteSecurityGroup{}, false},
-		{&egoscale.ListSecurityGroups{}, false},
-		{&egoscale.RevokeSecurityGroupEgress{}, false},
-		{&egoscale.RevokeSecurityGroupIngress{}, false},
-	},
-	"ssh": {
-		{&egoscale.RegisterSSHKeyPair{}, false},
-		{&egoscale.ListSSHKeyPairs{}, false},
-		{&egoscale.CreateSSHKeyPair{}, false},
-		{&egoscale.DeleteSSHKeyPair{}, false},
-		{&egoscale.ResetSSHKeyForVirtualMachine{}, false},
-	},
-	"affinity-group": {
-		{&egoscale.CreateAffinityGroup{}, false},
-		{&egoscale.DeleteAffinityGroup{}, false},
-		{&egoscale.ListAffinityGroups{}, false},
-		{&egoscale.UpdateVMAffinityGroup{}, false},
-	},
-	"vm group": {
-		{&egoscale.CreateInstanceGroup{}, false},
-		{&egoscale.ListInstanceGroups{}, false},
-	},
-	"tags": {
-		{&egoscale.CreateTags{}, false},
-		{&egoscale.DeleteTags{}, false},
-		{&egoscale.ListTags{}, false},
-	},
-	"nic": {
-		{&egoscale.ActivateIP6{}, false},
-		{&egoscale.AddIPToNic{}, false},
-		{&egoscale.ListNics{}, false},
-		{&egoscale.RemoveIPFromNic{}, false},
-	},
-	"address": {
-		{&egoscale.AssociateIPAddress{}, false},
-		{&egoscale.DisassociateIPAddress{}, false},
-		{&egoscale.ListPublicIPAddresses{}, false},
-		{&egoscale.UpdateIPAddress{}, false},
-	},
-	"async job": {
-		{&egoscale.QueryAsyncJobResult{}, false},
-		{&egoscale.ListAsyncJobs{}, false},
-	},
-	"apis": {
-		{&egoscale.ListAPIs{}, false},
-	},
-	"event": {
-		{&egoscale.ListEventTypes{}, false},
-		{&egoscale.ListEvents{}, false},
-	},
-	"offerings": {
-		{&egoscale.ListResourceDetails{}, false},
-		{&egoscale.ListResourceLimits{}, false},
-		{&egoscale.ListServiceOfferings{}, false},
-	},
-	"host": {
-		{&egoscale.ListHosts{}, true},
-		{&egoscale.UpdateHost{}, true},
-	},
-	"reversedns": {
-		{&egoscale.DeleteReverseDNSFromPublicIPAddress{}, false},
-		{&egoscale.QueryReverseDNSForPublicIPAddress{}, false},
-		{&egoscale.UpdateReverseDNSForPublicIPAddress{}, false},
-		{&egoscale.DeleteReverseDNSFromVirtualMachine{}, false},
-		{&egoscale.QueryReverseDNSForVirtualMachine{}, false},
-		{&egoscale.UpdateReverseDNSForVirtualMachine{}, false},
-	},
-}
-
-func printJSON(out string) {
-	if terminal.IsTerminal(syscall.Stdout) {
-		if _, err := fmt.Fprintln(os.Stdout, ""); err != nil {
-			log.Fatal(err)
-		}
-	} else {
-		if _, err := fmt.Fprintln(os.Stdout, out); err != nil {
-			log.Fatal(err)
-		}
-	}
-}
-
-func printResponseHelp(out io.Writer, response interface{}) {
-	value := reflect.ValueOf(response)
-	typeof := reflect.TypeOf(response)
-
-	w := tabwriter.NewWriter(out, 0, 0, 1, ' ', tabwriter.FilterHTML)
-	if _, err := fmt.Fprintln(w, "FIELD\tTYPE\tDOCUMENTATION"); err != nil {
-		log.Fatal(err)
-	}
-
-	for typeof.Kind() == reflect.Ptr {
-		typeof = typeof.Elem()
-		value = value.Elem()
-	}
-
-	for i := 0; i < typeof.NumField(); i++ {
-		field := typeof.Field(i)
-		tag := field.Tag
-		doc := "-"
-		if d, ok := tag.Lookup("doc"); ok {
-			doc = d
-		}
-
-		name := field.Type.Name()
-		if name == "" {
-			if field.Type.Kind() == reflect.Slice {
-				name = "[]" + field.Type.Elem().Name()
-			}
-		}
-
-		if json, ok := tag.Lookup("json"); ok {
-			n, _ := egoscale.ExtractJSONTag(field.Name, json)
-			if _, err := fmt.Fprintf(w, "%s\t%s\t%s\n", n, name, doc); err != nil {
-				log.Fatal(err)
-			}
-		}
-	}
-
-	if err := w.Flush(); err != nil {
-		log.Fatal(err)
 	}
 }
